requirements.txt
opencv-python
Pillow
numpy
pickle-mixin
os-sys
times